LAB1/consumers/preprocessor/requirements.txt
kafka-python==2.3.0
scikit-learn==1.7.2
pandas==2.2.2
numpy==2.0.2